Added Match activity
This adds the functionality of the matching.

To Brandon:
Check the three todo's in the MatchActivity.java
What you need to do is add firebase references so that the instance variables refer to the traits of each of the users. I also dont know how the chat activity works, so if possible, can you implement it so that the Connect button directs you to a chat with the current matched user.